refactor: refactored client code structure
