--- client/@types/global.d.ts
@@ -1,5 +1,0 @@
-export declare global {
-  interface Window {
-    stardust: { namedRoutes: Record<string, string> };
-  }
-}
