+++ client/index.ts
@@ -1,1 +1,17 @@
-export { initRoutes, stardust } from './Stardust';
+import { Stardust } from './Stardust';
+
+declare global {
+  interface Window {
+    stardust: { namedRoutes: Record<string, string> };
+  }
+}
+
+export let stardust;
+
+/**
+ * Initialize stardust
+ */
+export function initRoutes() {
+  const { namedRoutes } = window.stardust;
+  stardust = new Stardust(namedRoutes);
+}
