--- client/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2021",
-    "lib": ["ES2019", "DOM"],
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "declaration": true,
-    "baseUrl": "."
-  }
-}
